3ad8140 Merge branch '13257-SciberQuestToolkit-contribution-integration' of github.com:burlen/SciberQuestToolKit-ParaView-Dev into 13257-SciberQuestToolkit-contribution-integration
c4bc30d Include windows directory parsing functions in build.
2fb69bd Revert "Turn of the SciberQuestToolkit plugin"
7c1b4c6 Merge remote-tracking branch 'paraview/master' into 13257-SciberQuestToolkit-contribution-integration
8b0ae4e remove process monitor sources
830e15b warning cleanup and disable the process monitor

+++ Plugins/SciberQuestToolKit/BOVArrayImageIterator.h
@@ -21,7 +21,7 @@ step.
 class BOVArrayImageIterator
 {
 public:
-  BOVArrayImageIterator(const BOVTimeStepImage *step, int size)
+  BOVArrayImageIterator(const BOVTimeStepImage *step, size_t size)
        :
     Step(step),
     Idx(0),
@@ -36,7 +36,7 @@ public:
   /**
   Advance the iterator.
   */
-  virtual int Next()
+  virtual size_t Next()
     {
     if (this->Idx<this->End)
       {
@@ -66,8 +66,8 @@ private:
 
 protected:
   const BOVTimeStepImage *Step;
-  int Idx;
-  int End;
+  size_t Idx;
+  size_t End;
 };
 
 // Define some specific classes of the iterator
